requirement.txt
asgiref==3.2.7
Babel==2.8.0
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
curl-http==0.1.2
Django==1.11
django-money==1.1
django-prices==2.2.0
django-rest-framework==0.1.0
djangorestframework==3.6.2
enmerkar==0.7.1
httpie==2.1.0
idna==2.9
prices==1.0.0
psycopg2==2.8.5
py-moneyed==0.8.0
pycurl==7.43.0.5
Pygments==2.6.1
pytz==2020.1
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
typing==3.7.4.1
urllib3==1.25.9
